+++ python/taichi/__init__.py
@@ -1,8 +1,7 @@
 import sys
 
 from taichi._logging import *
-from taichi.core import get_os_name, package_root, require_version
-from taichi.core import ti_core as core
+from taichi.lib.core import ti_core as core
 from taichi.lang import *  # pylint: disable=W0622 # TODO(archibate): It's `taichi.lang.core` overriding `taichi.core`
 from taichi.main import main
 from taichi.tools import *
