+++ python/taichi/_testing.py
@@ -4,7 +4,7 @@ import itertools
 import os
 from tempfile import mkstemp
 
-from taichi.core import ti_core as _ti_core
+from taichi.lib.core import ti_core as _ti_core
 from taichi.lang import (cc, cpu, cuda, gpu, is_arch_supported, metal, opengl,
                          vulkan)
