+++ python/taichi/aot/record.py
@@ -1,6 +1,6 @@
 import os
 
-from taichi.core import ti_core
+from taichi.lib.core import ti_core
 
 
 def record_action_entry(name, contents):
