+++ python/taichi/lang/_ndarray.py
@@ -1,6 +1,6 @@
 import numpy as np
 import taichi.lang
-from taichi.core.util import ti_core as _ti_core
+from taichi.lib.core import ti_core as _ti_core
 from taichi.lang import impl
 from taichi.lang.enums import Layout
 from taichi.lang.util import (cook_dtype, has_pytorch, python_scope,
