+++ python/taichi/lang/expr.py
@@ -1,5 +1,5 @@
 import numpy as np
-from taichi.core.util import ti_core as _ti_core
+from taichi.lib.core import ti_core as _ti_core
 from taichi.lang import impl
 from taichi.lang.common_ops import TaichiOperations
 from taichi.lang.util import is_taichi_class
