+++ python/taichi/lang/field.py
@@ -1,5 +1,5 @@
 import taichi.lang
-from taichi.core.util import ti_core as _ti_core
+from taichi.lib.core import ti_core as _ti_core
 from taichi.lang.util import python_scope, to_numpy_type, to_pytorch_type
 
 import taichi as ti
