+++ python/taichi/lang/kernel_arguments.py
@@ -1,5 +1,5 @@
 import taichi.lang
-from taichi.core.util import ti_core as _ti_core
+from taichi.lib.core import ti_core as _ti_core
 from taichi.lang.any_array import AnyArray
 from taichi.lang.enums import Layout
 from taichi.lang.expr import Expr
