+++ python/taichi/lang/type_factory_impl.py
@@ -1,4 +1,4 @@
-from taichi.core.util import ti_core as _ti_core
+from taichi.lib.core import ti_core as _ti_core
 from taichi.lang import impl
