+++ python/taichi/lang/util.py
@@ -2,7 +2,7 @@ import functools
 import os
 
 import numpy as np
-from taichi.core.util import ti_core as _ti_core
+from taichi.lib.core import ti_core as _ti_core
 from taichi.lang import impl
 
 import taichi as ti
