+++ python/taichi/lib/core/__init__.py
@@ -1,3 +1,3 @@
-from taichi.core.util import *
+from taichi.lib.core.util import ti_core
 
 __all__ = [s for s in dir() if not s.startswith('_')]
