+++ python/taichi/profiler/kernelmetrics.py
@@ -1,4 +1,4 @@
-from taichi.core import ti_core as _ti_core
+from taichi.lib.core import ti_core as _ti_core
 
 
 class CuptiMetric:
