+++ python/taichi/snode/fields_builder.py
@@ -1,6 +1,6 @@
 from typing import Any, Optional, Sequence, Union
 
-from taichi.core.util import ti_core as _ti_core
+from taichi.lib.core import ti_core as _ti_core
 from taichi.lang import impl, snode
 from taichi.lang.exception import InvalidOperationError
 from taichi.snode.snode_tree import SNodeTree
