+++ python/taichi/ui/gui.py
@@ -4,7 +4,7 @@ import os
 
 import numpy as np
 import taichi.lang
-from taichi.core import ti_core as _ti_core
+from taichi.lib.core import ti_core as _ti_core
 from taichi.lang.field import Field, ScalarField
 from taichi.tools.util import core_veci, deprecated
