+++ python/taichi/ui/scene.py
@@ -1,4 +1,4 @@
-from taichi.core import ti_core as _ti_core
+from taichi.lib.core import ti_core as _ti_core
 from taichi.lang.impl import field
 from taichi.lang.kernel_impl import kernel
 from taichi.lang.matrix import Vector
