+++ python/taichi/ui/window.py
@@ -1,6 +1,6 @@
 import pathlib
 
-from taichi.core import ti_core as _ti_core
+from taichi.lib.core import ti_core as _ti_core
 from taichi.lang.impl import default_cfg
 
 from .canvas import Canvas
